sap-connector/requirements.txt
alembic==1.12.1
authlib==1.2.1
fastapi==0.104.1
httpx<0.25.0,>=0.24.0

itsdangerous==2.1.2
passlib[bcrypt]==1.7.4
prometheus_client==0.19.0
psycopg2-binary==2.9.9
pydantic==2.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
supabase==2.0.0
uvicorn==0.24.0
requests